% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{bbqr-package}
\alias{bbqr-package}
\title{The bbqr package is a bare-bones message queue in R}
\description{
The bbqr package is a bare-bones message queue in R
}
\examples{
# See ?bbqr for examples.
}
\references{
\url{https://github.com/wlandau/bbqr}
}
\author{
William Michael Landau \email{will.landau@gmail.com}
}
