% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtq.R
\docType{data}
\name{R6_txtq}
\alias{R6_txtq}
\title{R6 class for \code{txtq} objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
R6_txtq
}
\description{
See the \code{\link[=txtq]{txtq()}} function for full documentation and usage.
}
\seealso{
txtq
}
\keyword{datasets}
