% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{txtq-package}
\alias{txtq-package}
\title{The txtq package is a bare-bones message queue in R}
\description{
The txtq package is a bare-bones message queue in R
}
\examples{
# See ?txtq for examples.
}
\references{
\url{https://github.com/wlandau/txtq}
}
\author{
William Michael Landau \email{will.landau@gmail.com}
}
